+++ src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/entity/Product.kt
@@ -1,0 +1,22 @@
+package com.jeremiascortes.LoQueTengo.backend.entity
+
+import jakarta.persistence.*
+import org.hibernate.annotations.SQLDelete
+import org.hibernate.annotations.SQLRestriction
+
+@Entity
+@Table(name = "product")
+@SQLDelete(sql = "UPDATE product SET is_deleted = TRUE, deleted_at = now(), updated_at = now() WHERE id = ?")
+@SQLRestriction("is_deleted = false")
+class Product(
+
+    @Column(nullable = false, length = 255)
+    val name: String,
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "category", nullable = false)
+    var category: Category,
+
+    @Column(nullable = false, length = 255)
+    var barCode: String,
+) : BaseEntity()